pyproject.toml
[project]
name = "talentscout-hiring-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=4.0.0",
    "fastmcp>=3.1.1",
    "htbuilder>=0.9.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.11",
    "matplotlib>=3.10.8",
    "mongoengine>=0.29.3",
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
